Added afterImport callback for datatypes
