+++ src/Controllers/ImportController.php
@@ -86,11 +86,8 @@ class ImportController extends Base
                 $records = $dataType->getRecords();
                 try {
                     $this->module->$mapper->import($definitions[$dataTypeHandle], $records);
-
-                    // @TODO: Don't hardcode datatype in controller
-                    if ('fields' == $dataTypeHandle) {
-                        Craft::$app->fields->updateFieldVersion();
-                    }
+                    Schematic::info('Calling after import');
+                    $dataType->afterImport();
                 } catch (WrongEditionException $e) {
                     Schematic::error('Craft Pro is required for datatype '.$dataTypeHandle);
                 }
