+++ src/Converters/Models/CategoryGroup.php
@@ -39,8 +39,6 @@ class CategoryGroup extends Base
      */
     public function saveRecord(Model $record, array $definition): bool
     {
-        $this->resetCraftSiteServiceSiteIdsCache();
-
         return Craft::$app->categories->saveGroup($record);
     }
 
@@ -50,20 +48,5 @@ class CategoryGroup extends Base
     public function deleteRecord(Model $record): bool
     {
         return Craft::$app->categories->deleteGroupById($record->id);
-    }
-
-    /**
-     * Reset craft site service site ids cache using reflection.
-     */
-    private function resetCraftSiteServiceSiteIdsCache()
-    {
-        $obj = Craft::$app->sites;
-        $refObject = new \ReflectionObject($obj);
-        if ($refObject->hasProperty('_sitesById')) {
-            $refProperty1 = $refObject->getProperty('_sitesById');
-            $refProperty1->setAccessible(true);
-            $refProperty1->setValue($obj, null);
-            $obj->init(); // reload sites
-        }
     }
 }
