+++ src/Mappers/ElementIndexMapper.php
@@ -58,8 +58,6 @@ class ElementIndexMapper extends BaseComponent implements MapperInterface
      */
     public function import(array $settingDefinitions, array $elementTypes): array
     {
-        $this->resetCraftEditableSectionsCache();
-
         foreach ($settingDefinitions as $elementTypeName => $settings) {
             $elementType = 'craft\\elements\\'.$elementTypeName;
             $mappedSettings = $this->getMappedSettings($settings, 'handle', 'id');
@@ -103,19 +101,5 @@ class ElementIndexMapper extends BaseComponent implements MapperInterface
         }
 
         return $mappedSettings;
-    }
-
-    /**
-     * Reset craft editable sections cache using reflection.
-     */
-    private function resetCraftEditableSectionsCache()
-    {
-        $obj = Craft::$app->sections;
-        $refObject = new \ReflectionObject($obj);
-        if ($refObject->hasProperty('_editableSectionIds')) {
-            $refProperty1 = $refObject->getProperty('_editableSectionIds');
-            $refProperty1->setAccessible(true);
-            $refProperty1->setValue($obj, $obj->getAllSectionIds());
-        }
     }
 }
